v1-upgrade-generator/requirements.txt
certifi==2023.7.22
gitpython==3.1.24
semver==2.8.1
urllib3==1.26.18
